requirements.txt
key-generator
mysql_connector